refactor: add more exception interfaces and fix command queuer
